requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic[email]
sqlalchemy
openai
python-multipart
python-dotenv
asyncpg